Initialize Lagrange P2 space.

+++ mesh.py
@@ -1,7 +1,7 @@
 from typing import List
 import meshio
 import numpy as np
-from scipy.sparse import csr_matrix
+# from scipy.sparse import csr_matrix
 
 class Mesh:
     def load(self, mesh_name: str) -> None:
@@ -24,48 +24,17 @@ class Mesh:
         for cid in range(num_cells):
             cell = msh.cells[cid]
             if has_physical:
-                data = msh.cell_data["gmsh:physical"][cid].reshape(-1, 1) # reshape to a column vector
+                data = msh.cell_data["gmsh:physical"][cid].astype(np.uint32).reshape(-1, 1) # reshape to a column vector
             else:
                 data = np.zeros((cell.data.shape[0], 1), dtype=np.uint32)
             if cell.type == "vertex" and data != None:
                 self.point[cell.data, -1] = data
             elif cell.type == "line":
-                self.edge = np.vstack((self.edge, np.hstack((cell.data, data))))
+                self.edge = np.vstack((self.edge, np.hstack((cell.data.astype(np.uint32), data))))
             elif cell.type == "triangle":
-                self.tri = np.vstack((self.tri, np.hstack((cell.data, data))))
+                self.tri = np.vstack((self.tri, np.hstack((cell.data.astype(np.uint32), data))))
             else:
                 raise RuntimeError("Unrecognized cell type. ")
-            
-    # def isP2Mesh(self) -> bool:
-    #     return len(self.Np) == 2
-
-    # def buildP2Mesh(self) -> None:
-    #     """
-    #     Build the P2 mesh. 
-    #     """
-    #     assert(not self.isP2Mesh())
-    #     Np = self.Np[0]
-    #     T = self.tri
-    #     Nt = T.shape[0]
-    #     # figure out the new dofs
-    #     tri_edges = np.block([ [T[:,[0]], T[:,[1]]], [T[:,[2]], T[:,[3]]], [T[:,[2]], T[:,[0]]] ])
-    #     S = csr_matrix((np.ones((3*Nt,)), (np.min(tri_edges, axis=1), np.max(tri_edges, axis=1))), shape=(Np, Np))
-    #     S.data = np.arange(S.nnz, dtype=S.dtype) + (Np+1)
-    #     row, col = S.nonzero()
-    #     # build new nodes
-    #     new_points = 0.5 * (self.point[row, :] + self.point[col, :])
-    #     new_points[:, -1] = 0. # reset the tags
-    #     self.point = np.vstack((self.point, new_points))
-    #     self.Np += [self.point.shape[0]]
-    #     # update edges
-    #     self.edge = np.hstack((self.edge, self.edge[:, [-1]]))
-    #     self.edge[:,-2] = S[np.min(self.edge[:,0:2], axis=1), np.max(self.edge[:,0:2], axis=1)]
-    #     # update the triangles
-    #     self.tri = np.hstack((self.tri[:,0:3], np.zeros((Nt, 3), dtype=self.tri.dtype), self.tri[:,[-1]]))
-    #     self.tri[:,3] = S[np.min(self.tri[:,0:2], axis=1), np.max(self.tri[:,0:2], axis=1)]
-    #     self.tri[:,4] = S[np.min(self.tri[:,1:3], axis=1), np.max(self.tri[:,1:3], axis=1)]
-    #     self.tri[:,5] = S[np.min(self.tri[:,[0,2]], axis=1), np.max(self.tri[:,[0,2]], axis=1)]
-    #     # todo: sort nodes?
 
     def view(self, sub_dim: int, tags: List[int]) -> "Mesh":
         Np = self.point.shape[0]
@@ -101,5 +70,4 @@ class Mesh:
     def refine(self) -> "Mesh":
         fine_mesh = Mesh()
         return fine_mesh
-
